[PATCH] Time and seek optimizations
We did too many gettimeofday() calls, this patch cuts the number by
40%. Use clock_gettime() MONOTONIC instead, it is faster on my system
at least.

This patch also optimizes calling lseek() only when necessary for the
sync io engine.

+++ filesetup.c
@@ -434,7 +434,8 @@ void close_files(struct thread_data *td)
 	int i;
 
 	for_each_file(td, f, i) {
-		if (td->unlink && td->filetype == FIO_TYPE_FILE) {
+		if (!td->filename && td->unlink &&
+		    td->filetype == FIO_TYPE_FILE) {
 			unlink(f->file_name);
 			free(f->file_name);
 			f->file_name = NULL;
